feat: Complete UI/UX overhaul with 12+ new reusable components
- Enhanced CSS styling with WCAG AA accessibility compliance
- Added 12 new reusable React components (Loading, Modal, Alert, FormGroup, Table, Badge, etc.)
- Implemented full mobile responsive design
- Improved form validation with error/success states
- Added professional loading spinners and animations
- Created comprehensive component library with best practices
- Updated Dashboard with improved layouts and styling
- Added 4 detailed documentation guides for developers
- Better color contrast and visual hierarchy
- Touch-friendly buttons and interactive elements
- Smooth transitions and animations throughout app

+++ backend/src/routes/expenses.js
@@ -14,9 +14,9 @@ const initTable = async () => {
         description TEXT,
         amount DECIMAL(12,2) DEFAULT 0.00,
         vendor VARCHAR(255),
-        employee_id BIGINT UNSIGNED,
+        employee_id INT UNSIGNED,
         status ENUM('PENDING','APPROVED','REJECTED','PAID') DEFAULT 'PENDING',
-        approved_by BIGINT UNSIGNED,
+        approved_by INT UNSIGNED,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         FOREIGN KEY (employee_id) REFERENCES employees(id) ON DELETE SET NULL,
         FOREIGN KEY (approved_by) REFERENCES users(id) ON DELETE SET NULL
